requirements.txt
Django
asgiref
beautifulsoup4
black
bleach
click
django-ckeditor
django-js-asset
django-taggit
importlib-metadata
mypy-extensions
packaging
pathspec
Pillow
platformdirs
psycopg2
pycodestyle
Pygments
pyparsing
python-decouple
six
soupsieve
sqlparse
toml
tomli
tzdata
webencodings
whitenoise
zipp
django-tailwind